db: Add SQL console API
The idea is to allow users to edit their own data stored in database.
Although it still needs rate limits and other limits; and ability
to rollback / reset data if some unrecoverable mistake was done.

Also working with raw IDs is unpleasant, maybe mapping masters into database
(or actually storing them in database) would be a good solution, to be able, for example, do
[select m.member_id, mm.name, m.xp from user_members m left join master.members mm on m.id = mm.id]
in order to get members' XP along with their human-readable names.